% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats_names.R
\name{get_stats_names}
\alias{get_stats_names}
\title{Get Names of Statistics}
\usage{
get_stats_names()
}
\value{
a list of all the possible statistics in the dataset
}
\description{
Get Names of Statistics
}
\examples{
get_stats_names()
}
