% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bracketpredictr.R
\name{get_team_color}
\alias{get_team_color}
\title{Get the primary and secondary color for a team from the NCAA Colors data}
\usage{
get_team_color(team_name)
}
\arguments{
\item{team_name}{The name of the team to look up in the NCAA Colors data}
}
\value{
A vector containing the primary and secondary colors for the team
}
\description{
This function retrieves the primary and secondary colors for a specified team
from the NCAA Colors data.
}
\examples{
get_team_color("Utah St.")
}
\references{
The NCAA colors dataset used in this function was originally created by Luke
Benz for the ncaahoopsR package (https://github.com/lbenz730/ncaahoopR).
}
