% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_team_names.R
\name{get_team_names}
\alias{get_team_names}
\title{Get Team Names}
\usage{
get_team_names()
}
\value{
the names of all possible teams to use in analysis
}
\description{
Get Team Names
}
\examples{
get_team_names()
}
