% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_team_stats.R
\name{get_team_stats}
\alias{get_team_stats}
\title{The Function to Get Team Stats}
\usage{
get_team_stats(teams, stats = names(cbb), years = seq(2013, 2019))
}
\arguments{
\item{teams}{team or vector of two teams to get stats on}

\item{stats}{the specific statistics to be returned, default is all stats}

\item{years}{the year/years to return data for the teams}
}
\value{
a dataframe containing the stats for the team or teams, within the
year range specified
}
\description{
Returns a dataframe containing one or two teams stats. Which stats are
returned is specified by stats, and which year(s) they are returned for is
specified by the years parameter.
}
\examples{
get_team_stats("UCLA")
get_team_stats(c("Utah St.", "UNLV"),
  stats=c("CONF", "G", "W", "ADJOE"), years = 2017)
}
