% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bracketpredictr.R
\name{plot_stats}
\alias{plot_stats}
\title{Plot the desired statistics for two college basketball teams over time}
\usage{
plot_stats(team1, team2, stat)
}
\arguments{
\item{team1}{The name of the first team}

\item{team2}{The name of the second team}

\item{stat}{A vector of the statistics to plot for the teams
(e.g., c("W", "EFG_O", "EFG_D"))}
}
\value{
A ggplot object showing the statistics over time for the two teams
}
\description{
Plot the desired statistics for two college basketball teams over time
}
\examples{
plot_stats("Utah", "Utah St.", c("W", "EFG_O", "EFG_D"))
}
\references{
The NCAA colors dataset used in this function was originally created by Luke
Benz for the ncaahoopsR package (https://github.com/lbenz730/ncaahoopsR).
}
