% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Validate inputs}
\usage{
validate_inputs(teams, stats, years)
}
\arguments{
\item{teams}{The college basketball team you want to find the stats of}

\item{stats}{Different statistics provided about the basketball team entered}

\item{years}{The year that corresponds with the season you want to know the
stats of for the specific team}
}
\value{
If the validate function is called and its not valid it will return
a stop message
}
\description{
Validates inputs to make sure other functions are using correct arguments.
}
