% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_winner.R
\name{predict_winner}
\alias{predict_winner}
\title{Predict Winner of Two Teams}
\usage{
predict_winner(
  team1,
  team2,
  stats = names(cbb),
  years = 2019,
  show_stats = FALSE
)
}
\arguments{
\item{team1}{The first team playing in the game to be predicted}

\item{team2}{The second team playing in the game to be predicted}

\item{stats}{The specific statistics the user would like to be compared
between the two teams. Defaults to all available statistics.}

\item{years}{The year of stats the user desires to be compared, defaults to
2019 which is the most recent year.}

\item{show_stats}{If true will return the table with the stats for the user
to view}
}
\value{
a print statement with a predicted winner, and stats if requested
}
\description{
This function takes two college basketball teams, compares their statistics
for a given year, and predicts the winner based on which team has higher
values in the majority of the chosen statistics. If there is a tie, the
function will just print "Tie". The function allows users to
specify a particular set of statistics and the year for the comparison.
Optionally, it can also print the statistics for both teams (before printing
the winner) if show_stats is set to TRUE.
}
\examples{
predict_winner("Utah St.", "Utah Valley", show_stats = TRUE)
predict_winner("Utah St.", "Utah", c("G", "W", "ADJOE", "ADJDE",
"X3P_O", "X3P_D", "ORB"))
}
